companion method
